requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
supabase==2.6.0
openai==1.40.0
python-dotenv==1.0.1
